requirements.txt
django
djangorestframework
django-cors-headers
user-agents
pillow
pyotp
